+++ vite-plugin-ssr/__internal/setup.d.ts
@@ -1,2 +1,2 @@
 // Help TS's resolver until it supports `package.json#exports`
-export * from '../dist/cjs/node/page-files/setup'
+export * from '../dist/cjs/node/runtime/page-files/setup'
