+++ vite-plugin-ssr/__internal/setup.js
@@ -1,2 +1,2 @@
 // eslint doesn't seem to support `package.json#exports`.
-module.exports = require('../dist/cjs/node/page-files/setup')
+module.exports = require('../dist/cjs/node/runtime/page-files/setup')
