+++ vite-plugin-ssr/node/runtime/globalContext.ts
