+++ vite-plugin-ssr/node/runtime/html/injectAssets.ts
