+++ vite-plugin-ssr/node/runtime/html/injectAssets/injectAssets__public.ts
