+++ vite-plugin-ssr/node/runtime/html/injectAssets/sanitizeJson.ts
