+++ vite-plugin-ssr/node/runtime/html/stream/react-streaming.ts
