+++ vite-plugin-ssr/node/runtime/renderPage/assertArguments.ts
