+++ vite-plugin-ssr/node/runtime/renderPage/getPageAssets.ts
