+++ vite-plugin-ssr/node/runtime/renderPage/logError.ts
